[fix]응답 객체 구체화 (#42)
응답 객체를 상속을 이용해 projection을 구체화하여 불필요한 field 응답을 제한
